2014/03/django-rest-framework/src/requirements.txt
django==1.6
djangorestframework==2.3.13
django-debug-toolbar==1.0.1
south==0.8.4
django-wkhtmltopdf==1.2.2
Pillow==2.3.1
easy_thumbnails==1.5
